Remove CLIs, split up league from coordinator
CLIs are too much of external notions to be at this level. Instead, the CLIs can be broken out into another repository which can use this core library.

The Coordinator is also too heavy in its current form and it should be decoupled as much as it can from the League Aggregate